Added Lockable Rooms
Rooms now have a boolean indicating whether they are locked or not.
If this boolean is true for a room, the player cannot go to said room.
Also added a LevelHandler, that unlocks levels as the player progresses based on a condition, that is, a number of waste, the player has to have sorted before a room unlocks.